-refactor : 회원가입 인증 서비스 수정
